feat(blog): extract logic to a custom hook for better reusability
