+++ app/blog/page.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useState, useEffect } from 'react';
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Calendar, ArrowRight } from "lucide-react";
@@ -8,68 +7,16 @@ import Link from "next/link";
 import Image from "next/image";
 import { Navbar } from "../components/navbar";
 import { Footer } from "../components/footer";
-
-interface BlogPost {
-  id: number;
-  title: string;
-  content: string;
-  imageUrl: string;
-  imagePublicId: string;
-  createdAt: string;
-}
+import { usePostsList } from "../hooks/usePostsList";
 
 export default function BlogPage() {
-  const [blogPosts, setBlogPosts] = useState<BlogPost[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    const fetchPosts = async () => {
-      try {
-        setLoading(true);
-        setError(null);
-        
-        const response = await fetch('/api/posts', {
-          headers: {
-            'Content-Type': 'application/json'
-          }
-        });
-        
-        if (!response.ok) {
-          const errorText = await response.text();
-          throw new Error(`Error ${response.status}: ${errorText || 'Error desconocido'}`);
-        }
-        
-        const data = await response.json();
-        setBlogPosts(data);
-      } catch (err) {
-        console.error('Error fetching posts:', err);
-        setError(err instanceof Error ? err.message : 'Error al cargar los posts');
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchPosts();
-  }, []);
-
-  const handleRetry = () => {
-    setError(null);
-    setLoading(true);
-  };
-
-  const formatDate = (dateString: string) => {
-    try {
-      const date = new Date(dateString);
-      return date.toLocaleDateString('es-ES', {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric'
-      });
-    } catch {
-      return dateString;
-    }
-  };
+  const {
+    blogPosts,
+    loading,
+    error,
+    handleRetry,
+    formatDate
+  } = usePostsList();
 
   if (loading) {
     return (
