+++ app/hooks/usePostsList.tsx
@@ -1,0 +1,75 @@
+'use client';
+
+import { useState, useEffect } from 'react';
+
+interface BlogPost {
+  id: number;
+  title: string;
+  content: string;
+  imageUrl: string;
+  imagePublicId: string;
+  createdAt: string;
+}
+
+export function usePostsList() {
+  const [blogPosts, setBlogPosts] = useState<BlogPost[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  const fetchPosts = async () => {
+    try {
+      setLoading(true);
+      setError(null);
+      
+      const response = await fetch('/api/posts', {
+        headers: {
+          'Content-Type': 'application/json'
+        }
+      });
+      
+      if (!response.ok) {
+        const errorText = await response.text();
+        throw new Error(`Error ${response.status}: ${errorText || 'Error desconocido'}`);
+      }
+      
+      const data = await response.json();
+      setBlogPosts(data);
+    } catch (err) {
+      console.error('Error fetching posts:', err);
+      setError(err instanceof Error ? err.message : 'Error al cargar los posts');
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const handleRetry = () => {
+    setError(null);
+    setLoading(true);
+    fetchPosts();
+  };
+
+  const formatDate = (dateString: string) => {
+    try {
+      const date = new Date(dateString);
+      return date.toLocaleDateString('es-ES', {
+        year: 'numeric',
+        month: 'long',
+        day: 'numeric'
+      });
+    } catch {
+      return dateString;
+    }
+  };
+
+  useEffect(() => {
+    fetchPosts();
+  }, []);
+
+  return {
+    blogPosts,
+    loading,
+    error,
+    handleRetry,
+    formatDate
+  };
+}
